+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## [Unreleased]
 
+## [0.17.0] - 2023-04-12
+
 ### Added
 
 - Added dashboard widget